fix: restore support for 32-bit platforms (#128)
PR #127 addressed a gosec-triggered warning and introduced additional
checks on salt length and hash length, preventing them from being
greater than MaxUint32. These checks prevent the library from being
compiled on 32-bit platforms, since, on these platforms, salt length
and hash length are 32-bit integers and cannot be compared to MaxUint32
due to an overflow. This patch fixes the issue by casting ints as
uint64s.

+++ encoding.go
@@ -270,7 +270,7 @@ func Decode(encoded []byte) (Raw, error) {
 	hash := make([]byte, enc64.DecodedLen(len(h)))
 	sl, se := enc64.Decode(salt, s)
 	hl, he := enc64.Decode(hash, h)
-	if se != nil || he != nil || sl <= 0 || hl <= 0 || sl > math.MaxUint32 || hl > math.MaxUint32 {
+	if se != nil || he != nil || sl <= 0 || hl <= 0 || uint64(sl) > math.MaxUint32 || uint64(hl) > math.MaxUint32 {
 		return Raw{}, ErrDecodingFail
 	}
